Optimize Array in da.from_zarr
This improves task fusion.
