+++ dask/array/core.py
@@ -2861,7 +2861,9 @@ def from_zarr(
     chunks = chunks if chunks is not None else z.chunks
     if name is None:
         name = "from-zarr-" + tokenize(z, component, storage_options, chunks, **kwargs)
-    return from_array(z, chunks, name=name)
+    # We inline the zarr.Array objects.
+    arr = from_array(z, chunks, name=name, getitem=getter_inline)
+    return arr
 
 
 def to_zarr(
